Performance improvement parity with bacom
